requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
filelock==3.13.1
Flask==2.3.3
flask-restx==1.2.0
Flask-Testing==0.8.1
fsspec==2023.10.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.19.4
hyperframe==5.2.0
idna==2.10
img2vec-pytorch==1.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
langdetect==1.0.9
lit==17.0.5
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.35.2
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==2.3.8
